src/requirements.txt
# Core pipeline
pydantic>=2.0
pandas>=2.0
openpyxl>=3.1
httpx>=0.27
tenacity>=8.0
python-dotenv>=1.0
pyyaml>=6.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
